fix(catalog-popularity-freetext-resolve): isolate BS#1822 integration tests in their own spec file
The BS#1822 play-floor/drain-order tests were appended as a second describe
block in the BS#1767 enumerate spec, sharing that file's module registry.
That spec's afterAll closes @wxyc/database's shared postgres-js pool, so the
second describe block's queries hit an already-closed connection
(CONNECTION_ENDED) once the first describe finished - all 4 new tests failed
in CI. Moving them to their own spec file gives them an independent
@wxyc/database import (Jest resets the module registry per test file), so
the pool is untouched by the other spec's teardown.
